test: drop error codes from test output
In the newest Tarantool the error objects are encoded as MP_ERROR
in netbox connections by default. This allows to transfer their
full data including correct error code, even when returned as
nil,err instead of throwing the error as an exception.

Because of that some vshard tests failed - they expected the error
codes returned from the server to be wrong.

On the other hand, the codes can't be fixed, because then the
tests wouldn't work on 1.10.

The patch just drops the codes. At least while need to support
Tarantool versions not using MP_ERROR in netbox by default.